% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cvode}
\alias{cvode}
\title{cvode}
\usage{
cvode(time_vec, IC, xpsexp, reltolerance, abstolerance)
}
\arguments{
\item{time_vec}{time vector}

\item{IC}{Initial Conditions}

\item{xpsexp}{External pointer to RHS function}

\item{reltolerance}{Relative Tolerance (a scalar)}

\item{abstolerance}{Absolute Tolerance (a vector with length equal to ydot)}
}
\description{
CVODE solver to solve stiff ODEs
}
