% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cvodes}
\alias{cvodes}
\title{cvodes}
\usage{
cvodes(time_vector, IC, input_function, Parameters, reltolerance,
  abstolerance, SensType = "STG", ErrCon = "F")
}
\arguments{
\item{time_vector}{time vector}

\item{IC}{Initial Conditions}

\item{input_function}{Right Hand Side function of ODEs}

\item{Parameters}{Parameters input to ODEs}

\item{reltolerance}{Relative Tolerance (a scalar)}

\item{abstolerance}{Absolute Tolerance (a vector with length equal to ydot)}

\item{SensType}{Sensitivity Type - allowed values are Staggered (default)", "STG" (for Staggered) or "SIM" (for Simultaneous)}

\item{ErrCon}{Error Control - allowed values are TRUE or FALSE (default)
//'@example /inst/examples/cvs_Roberts_dns.r}
}
\description{
CVODES solver to solve ODEs and calculate sensitivities
}
