% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ida}
\alias{ida}
\title{ida}
\usage{
ida(
  time_vector,
  IC,
  IRes,
  input_function,
  Parameters,
  reltolerance = 1e-04,
  abstolerance = 1e-04
)
}
\arguments{
\item{time_vector}{time vector}

\item{IC}{Initial Value of y}

\item{IRes}{Inital Value of ydot}

\item{input_function}{Right Hand Side function of DAEs}

\item{Parameters}{Parameters input to ODEs}

\item{reltolerance}{Relative Tolerance (a scalar, default value  = 1e-04)}

\item{abstolerance}{Absolute Tolerance (a scalar or vector with length equal to ydot, default = 1e-04)}
}
\description{
IDA solver to solve stiff DAEs
}
